+++ packages/api-shared/src/v1/index.ts
@@ -13,3 +13,4 @@
  */
 export * as auth from "./auth";
 export * as common from "./common";
+export * as users from "./users";